Add LabelEditor component Handles events fired by active annotations. Currently just deletes annotations.
